DiSSOmniag add RPC support, DiSSOmniaG_live added update support

+++ DiSSOmniaG/dissomniag/cliApi/ManageUsers.py
@@ -292,6 +292,9 @@ class delUser(CliMethodABCClass.CliMethodABCClass):
         
         try:
             user = session.query(User).filter(User.username == options.username).one()
+            if user.isMaintain:
+                self.printError("Permission denied: You cannot delete a maintainance User!")
+                return
             session.delete(user)
             session.commit()
             session.flush()
